feat: Extract login logic to custom hook and handle errors in component
This commit extracts the login logic previously within the GenderDuelPage component into a custom hook, useUserLogin. It handles the state and side effects related to the login process, including validation and dispatching actions to Redux.

This commit extracts the check of user exitence and guest login logic into the useUserUsername hook.

Errors are now thrown from the hook and caught in the GenderDuelPage component, where they trigger a toast notification to the user. This keeps UI logic and data logic separate, improving code organization and reusability.