feat: Add native language code to course data
Include 'native_language_code' field in course data with value 'en' to
indicate the native language of the course. This enhances the ability to
categorize and filter courses based on the language they are originally
created in. It ensures a better user experience when searching for
courses in their native language.